Converted 'combine' command
- Moved the combine CLI into the new CombineCmd class.
- Added ability to specify a 'format' in the KsconfCmd class.
- Added ordered results to the entrypoint loader so that subcommand order can
  be preserved.
